Refactor events indexers
